+++ projects/angular2-smart-table/src/lib/components/filter/filter.module.ts
@@ -1,6 +1,5 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 
 import {FilterComponent} from './filter.component';
 import {DefaultFilterComponent} from "./default-filter.component";
@@ -25,8 +24,6 @@ const FILTER_COMPONENTS = [
 @NgModule({
   imports: [
     CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
   ],
   declarations: [
     ...FILTER_COMPONENTS,