+++ projects/angular2-smart-table/src/lib/components/pager/pager.module.ts
@@ -1,13 +1,11 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
 
 import {PagerComponent} from './pager.component';
 
 @NgModule({
   imports: [
     CommonModule,
-    FormsModule,
   ],
   declarations: [
     PagerComponent,