+++ projects/demo/src/app/pages/examples/custom-edit-view/custom-filter.component.ts
@@ -1,44 +1,20 @@
-import {Component, OnChanges, OnInit, SimpleChanges} from '@angular/core';
-import {FormControl} from '@angular/forms';
-import {debounceTime, distinctUntilChanged} from 'rxjs/operators';
+import {Component} from '@angular/core';
 
 import {DefaultFilter} from 'angular2-smart-table';
 
 @Component({
     template: `
     <input
-      #number
+      #textfield
       [ngClass]="inputClass"
-      [formControl]="inputControl"
-      class="form-control"
       [placeholder]="column.title"
-      type="number">
+      type="number"
+      [value]="query"
+      (change)="onValueChanged(textfield.value)"
+      (keyup)="onValueChanged(textfield.value)"
+    >
   `,
     standalone: false
 })
-export class CustomFilterComponent extends DefaultFilter implements OnInit, OnChanges {
-  inputControl = new FormControl();
-
-  constructor() {
-    super();
-  }
-
-  ngOnInit() {
-    this.inputControl.valueChanges
-      .pipe(
-        distinctUntilChanged(),
-        debounceTime(this.debounceTime),
-      )
-      .subscribe(value => {
-        this.query = value;
-        this.setFilter();
-      });
-  }
-
-  ngOnChanges(changes: SimpleChanges) {
-    if (changes.query) {
-      this.query = changes.query.currentValue;
-      this.inputControl.setValue(this.query);
-    }
-  }
+export class CustomFilterComponent extends DefaultFilter {
 }