+++ projects/angular2-smart-table/CHANGELOG.md
@@ -7,6 +7,7 @@ This document lists the changes introduced by this fork.
 * Upgrade to Angular 19
 * Add new `filter.debounceTime` setting
 * Fix broken backwards compatibility with rxjs ^6.5.3
+* Remove dependency to `@angular/forms`
 
 ## Version 3.4.0